remove response_handler
To show to the user that you can create
custome Handler for your CUrl Builder
